+++ lib/main.dart
@@ -70,26 +70,7 @@ void main() async {
 
   log.info('Setup bloc oberserver');
   Bloc.observer = GenericBlocObserver();
-
-  Directory legacyAppDataDir = await getApplicationCacheDirectory();
   Directory appDataDir = await getApplicationDocumentsDirectory();
-  // @todo: Remove later!
-  if (File(path.join(legacyAppDataDir.path, 'data.db')).existsSync()) {
-    log.info(
-        'Legacy application data found in cache directory. Migrate data to new path.');
-    for (FileSystemEntity f in legacyAppDataDir.listSync()) {
-      String filename = f.uri.pathSegments.last;
-      String newFullFilename = path.join(appDataDir.path, filename);
-      if (!await File(newFullFilename).exists()) {
-        log.info('Moving file $f to $newFullFilename');
-        await (f as File).copy(newFullFilename);
-        await f.delete();
-      }
-    }
-  } else {
-    log.info(
-        'No application data found in cache directory. Using documents directory.');
-  }
 
   log.info('Run app');
   runApp(