+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Improvement of the database handling
 
+### Removed
+
+- Deprecated code for backwards compatibility
+
 ## [0.13.0] - 2025-05-22
 
 ### Changed